# Integrate JGroups cluster peer enhancements.