--- a/actors/json/src/main/java/cloud/orbit/actors/extensions/json/InMemoryJSONStorageExtension.java
+++ b/actors/json/src/main/java/cloud/orbit/actors/extensions/json/InMemoryJSONStorageExtension.java
@@ -48,6 +48,7 @@ public class InMemoryJSONStorageExtension implements StorageExtension
     private ObjectMapper mapper = new ObjectMapper();
     private String name;
 
+    @Override
     public String getName()
     {
         return name;