--- a/actors/runtime/src/main/java/cloud/orbit/actors/runtime/Hosting.java
+++ b/actors/runtime/src/main/java/cloud/orbit/actors/runtime/Hosting.java
@@ -515,7 +515,7 @@ public class Hosting implements NodeCapabilities, Startable, PipelineExtension
         {
             try
             {
-                serverNodesUpdateMutex.wait(100);
+                serverNodesUpdateMutex.wait(250);
             }
             catch (InterruptedException e)
             {