```text
old / new with atomized executor

Benchmark                                         Mode  Cnt      Score     Error  Units
SingleNodeBenchmark.requestThroughput            thrpt  100  63069.039 ± 829.549  ops/s
SingleNodeBenchmark.avgRequestTime_batched        avgt  100     13.701 ±   0.219  us/op
SingleNodeBenchmark.avgRequestTime_singleThread   avgt  100     30.104 ±   0.290  us/op


new with single executor

Benchmark                                         Mode  Cnt       Score      Error  Units
SingleNodeBenchmark.requestThroughput            thrpt  100  190695.723 ± 1625.727  ops/s
SingleNodeBenchmark.avgRequestTime_batched        avgt  100       5.123 ±    0.039  us/op
SingleNodeBenchmark.avgRequestTime_singleThread   avgt  100      35.951 ±    0.257  us/op


new with second executor

Benchmark                                         Mode  Cnt       Score      Error  Units
SingleNodeBenchmark.requestThroughput            thrpt  100  173215.276 ± 1993.871  ops/s
SingleNodeBenchmark.avgRequestTime_batched        avgt  100       6.878 ±    0.188  us/op
SingleNodeBenchmark.avgRequestTime_singleThread   avgt  100      35.909 ±    0.292  us/op
```