@johnou 2017-06-30
Not sure what's up with the hit in batched..

```text
Before
Benchmark                                         Mode  Cnt      Score     Error  Units
SingleNodeBenchmark.requestThroughput            thrpt  100  63033.949 ± 646.801  ops/s
SingleNodeBenchmark.avgRequestTime_batched        avgt  100     13.905 ±   0.170  us/op
SingleNodeBenchmark.avgRequestTime_singleThread   avgt  100     30.291 ±   0.281  us/op

After
Benchmark                                         Mode  Cnt       Score      Error  Units
SingleNodeBenchmark.requestThroughput            thrpt  100  192216.205 ± 2304.724  ops/s
SingleNodeBenchmark.avgRequestTime_batched        avgt  100       5.375 ±    0.141  us/op
SingleNodeBenchmark.avgRequestTime_singleThread   avgt  100      37.554 ±    0.971  us/op
```