@johnou 2017-06-30
After a bit of digging batched suffers a performance hit when using the same executor service as execution, by using a different executor batched score is on par with the previous score.