@johnou 2017-07-12
Infinispan release with Jgroups fix should be cut this week, also requires further benchmarking.